api/files/api/requirements.txt
Flask-Cors==2.1.2
Flask-Migrate==1.8.0
Flask-SQLAlchemy==2.1
Flask-Script==2.0.5
Flask==0.10.1
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
MySQL-python==1.2.5
Paste==2.0.3
PyYAML==3.11
SQLAlchemy-Utils==0.32.0
SQLAlchemy==1.0.12
Werkzeug==0.11.5
alembic==0.8.6
amqp==1.4.9
anyjson==0.3.3
apispec==0.9.1
argh==0.26.1
billiard==3.3.0.23
blinker==1.4
boto3==1.4.6
botocore==1.6.7
celery==3.1.23
cffi==1.5.2
cryptography==1.3.1
docutils==0.12
elasticsearch-dsl==2.1.0
elasticsearch==2.4.0
enum34==1.1.2
futures==3.0.5
google-api-python-client==1.5.0
gunicorn==19.4.5
idna==2.1
ipaddress==1.0.16
itsdangerous==0.24
jmespath==0.9.0
kombu==3.0.35
marshmallow==2.6.1
ndg-httpsclient==0.4.0
numpy==1.11.0
pathtools==0.1.2
pyOpenSSL==16.0.0
pyasn1==0.1.9
pycparser==2.14
python-dateutil==2.5.3
python-editor==1.0
pytz==2016.3
redis==2.10.5
requests==2.9.1
requests-aws4auth==0.9
scipy==0.17.1
statsd==3.2.1
url==0.2.0
urllib3==1.14
watchdog==0.8.3
azure==2.0.0rc6
azure-mgmt-commerce==0.30.0rc6
newrelic==2.88.1.73
